test: replace `beforeEach` by `before` hooks across LSP test suites (#534)
* test: replace `beforeEach` -> `before` in LSP11

* test: replace `beforeEach` -> `before` in LSP9

* test: replace `beforeEach` -> `before` in Mocks

* test: replace `beforeEach` -> `before` in LSP4 tests

* test: replace `beforeEach` -> `before` in LSP7 tests

* test: replace `beforeEach` -> `before` in LSP8 tests

* test: replace `beforeEach` -> `before` in LSP17 tests

* test: replace `beforeEach` -> `before` in LSP20 interactions tests

* test: replace `beforeEach` -> `before` for LSP14 `renounceOwnership` test + improve tests flow

* test: replace `beforeEach` -> `before` in LSP20/LSP6 OWNER permission tests

* test: replace `beforeEach` -> `before` in LSP20 CHANGEOWNER tests

* test: move tests for setAllowedCalls/ERC725YDataKeys in their own file in LSP20

* test: replace `beforeEach` -> `before` in LSP20 -> ADD/EDIT permissions tests

* test: replace `beforeEach` -> `before` in LSP20WithLSP14 tests
