+++ tests/LSP17ContractExtension/LSP17Extendable.behaviour.ts
@@ -1,13 +1,11 @@
 import { expect } from "chai";
 import { ethers } from "hardhat";
-
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
+
 import {
   LSP0ERC725Account,
   LSP9Vault,
   CheckerExtension__factory,
-  CheckerExtension,
-  RevertStringExtension,
   ERC165Extension,
   ERC165Extension__factory,
   RevertStringExtension__factory,
@@ -31,7 +29,9 @@ import {
   RevertFallbackExtension__factory,
 } from "../../types";
 
+// helpers
 import { abiCoder, provider } from "../utils/helpers";
+
 // constants
 import { ERC725YDataKeys } from "../../constants";
 
@@ -307,7 +307,7 @@ export const shouldBehaveLikeLSP17 = (
             });
 
             describe("when the extension is set", () => {
-              beforeEach(async () => {
+              before(async () => {
                 const checkerExtension = await new CheckerExtension__factory(
                   context.accounts[0]
                 ).deploy();
@@ -383,7 +383,7 @@ export const shouldBehaveLikeLSP17 = (
         });
 
         describe("when calling an extension that reverts with string error", () => {
-          beforeEach(async () => {
+          before(async () => {
             const revertStringExtension =
               await new RevertStringExtension__factory(
                 context.accounts[0]
@@ -416,7 +416,8 @@ export const shouldBehaveLikeLSP17 = (
 
         describe("when calling an extension that reverts with Custom error with tx.origin and msg.sender as parameters", () => {
           let revertCustomExtension: RevertCustomExtension;
-          beforeEach(async () => {
+
+          before(async () => {
             revertCustomExtension = await new RevertCustomExtension__factory(
               context.accounts[0]
             ).deploy();
@@ -449,7 +450,8 @@ export const shouldBehaveLikeLSP17 = (
 
         describe("when calling an extension that emits an event", () => {
           let emitEventExtension: EmitEventExtension;
-          beforeEach(async () => {
+
+          before(async () => {
             emitEventExtension = await new EmitEventExtension__factory(
               context.accounts[0]
             ).deploy();
@@ -475,7 +477,8 @@ export const shouldBehaveLikeLSP17 = (
 
         describe("when calling an extension that returns a string", () => {
           let nameExtension: NameExtension;
-          beforeEach(async () => {
+
+          before(async () => {
             nameExtension = await new NameExtension__factory(
               context.accounts[0]
             ).deploy();
@@ -503,7 +506,8 @@ export const shouldBehaveLikeLSP17 = (
 
         describe("when calling an extension that returns a number", () => {
           let ageExtension: AgeExtension;
-          beforeEach(async () => {
+
+          before(async () => {
             ageExtension = await new AgeExtension__factory(
               context.accounts[0]
             ).deploy();
@@ -529,7 +533,8 @@ export const shouldBehaveLikeLSP17 = (
 
         describe("when calling an extension that modify the state of the extension", () => {
           let transferExtension: TransferExtension;
-          beforeEach(async () => {
+
+          before(async () => {
             transferExtension = await new TransferExtension__factory(
               context.accounts[0]
             ).deploy();
@@ -570,7 +575,8 @@ export const shouldBehaveLikeLSP17 = (
 
         describe("when calling an extension that reenter the fallback function of the account", () => {
           let reenterAccountExtension: ReenterAccountExtension;
-          beforeEach(async () => {
+
+          before(async () => {
             reenterAccountExtension =
               await new ReenterAccountExtension__factory(
                 context.accounts[0]
@@ -602,7 +608,8 @@ export const shouldBehaveLikeLSP17 = (
           describe("when reentering with a call to an extension that emits an event", () => {
             describe("when reentering before setting the extension", () => {
               let emitEventExtension: EmitEventExtension;
-              beforeEach(async () => {
+
+              before(async () => {
                 emitEventExtension = await new EmitEventExtension__factory(
                   context.accounts[0]
                 ).deploy();
@@ -625,7 +632,8 @@ export const shouldBehaveLikeLSP17 = (
             });
             describe("when reentering after setting the extension", () => {
               let emitEventExtension: EmitEventExtension;
-              beforeEach(async () => {
+
+              before(async () => {
                 emitEventExtension = await new EmitEventExtension__factory(
                   context.accounts[0]
                 ).deploy();
@@ -691,7 +699,8 @@ export const shouldBehaveLikeLSP17 = (
     describe("when calling with calldata that is not checked for extension", () => {
       describe("when calling with a payload of length less than 4bytes", () => {
         let revertFallbackExtension: RevertFallbackExtension;
-        beforeEach(async () => {
+
+        before(async () => {
           revertFallbackExtension = await new RevertFallbackExtension__factory(
             context.accounts[0]
           ).deploy();
@@ -794,7 +803,8 @@ export const shouldBehaveLikeLSP17 = (
         describe("when there is an extension set for bytes4(0)", () => {
           describe("when setting an extension that reverts", () => {
             let revertFallbackExtension: RevertFallbackExtension;
-            beforeEach(async () => {
+
+            before(async () => {
               revertFallbackExtension =
                 await new RevertFallbackExtension__factory(
                   context.accounts[0]
@@ -850,11 +860,13 @@ export const shouldBehaveLikeLSP17 = (
     describe("use cases", async () => {
       describe("when interacting with a contract that require the recipient to implement onERC721Received function to mint", () => {
         let token: RequireCallbackToken;
-        beforeEach(async () => {
+
+        before(async () => {
           token = await new RequireCallbackToken__factory(
             context.accounts[0]
           ).deploy();
         });
+
         describe("when minitng to the account", () => {
           describe("before setting the onERC721ReceivedExtension", () => {
             it("should fail since onERC721Received is not implemented", async () => {
@@ -864,7 +876,8 @@ export const shouldBehaveLikeLSP17 = (
 
           describe("after setting the onERC721ReceivedExtension", () => {
             let onERC721ReceivedExtension: OnERC721ReceivedExtension;
-            beforeEach(async () => {
+
+            before(async () => {
               onERC721ReceivedExtension =
                 await new OnERC721ReceivedExtension__factory(
                   context.accounts[0]
