+++ tests/LSP7DigitalAsset/extensions/LSP7CappedSupply.test.ts
@@ -40,7 +40,7 @@ describe("LSP7CappedSupply", () => {
     describe("when deploying the contract", () => {
       let context: LSP7CappedSupplyTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -96,7 +96,7 @@ describe("LSP7CappedSupply", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP7CappedSupplyTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -115,8 +115,6 @@ describe("LSP7CappedSupply", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
