+++ tests/LSP7DigitalAsset/extensions/LSP7CompatibleERC20.behaviour.ts
@@ -1283,7 +1283,7 @@ export const shouldInitializeLikeLSP7CompatibleERC20 = (
 ) => {
   let context: LSP7InitializeTestContext;
 
-  beforeEach(async () => {
+  before(async () => {
     context = await buildContext();
   });
