+++ tests/LSP7DigitalAsset/extensions/LSP7CompatibleERC20.test.ts
@@ -46,7 +46,7 @@ describe("LSP7CompatibleERC20", () => {
     describe("when deploying the contract", () => {
       let context: LSP7CompatibleERC20TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -148,7 +148,7 @@ describe("LSP7CompatibleERC20", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP7CompatibleERC20TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -167,8 +167,6 @@ describe("LSP7CompatibleERC20", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
