+++ tests/LSP7DigitalAsset/extensions/LSP7Mintable.test.ts
@@ -44,7 +44,7 @@ describe("LSP7Mintable", () => {
     describe("when deploying the contract", () => {
       let context: LSP7MintableTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -120,7 +120,7 @@ describe("LSP7Mintable", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP7MintableTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -139,8 +139,6 @@ describe("LSP7Mintable", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
