+++ tests/LSP8IdentifiableDigitalAsset/LSP8IdentifiableDigitalAsset.test.ts
@@ -77,7 +77,7 @@ describe("LSP8", () => {
       describe("once the contract was deployed", () => {
         let context: LSP8TestContext;
 
-        beforeEach(async () => {
+        before(async () => {
           context = await buildTestContext();
         });
 
@@ -149,7 +149,7 @@ describe("LSP8", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP8TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -178,8 +178,6 @@ describe("LSP8", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
