+++ tests/LSP8IdentifiableDigitalAsset/extensions/LSP8CappedSupply.test.ts
@@ -39,7 +39,7 @@ describe("LSP8CappedSupply", () => {
     describe("when deploying the contract", () => {
       let context: LSP8CappedSupplyTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -95,7 +95,7 @@ describe("LSP8CappedSupply", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP8CappedSupplyTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -114,8 +114,6 @@ describe("LSP8CappedSupply", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
