+++ tests/LSP8IdentifiableDigitalAsset/extensions/LSP8CompatibleERC721.behaviour.ts
@@ -1353,7 +1353,7 @@ export const shouldInitializeLikeLSP8CompatibleERC721 = (
 ) => {
   let context: LSP8InitializeTestContext;
 
-  beforeEach(async () => {
+  before(async () => {
     context = await buildContext();
   });
