+++ tests/LSP8IdentifiableDigitalAsset/extensions/LSP8CompatibleERC721.test.ts
@@ -55,7 +55,7 @@ describe("LSP8CompatibleERC721", () => {
     describe("when deploying the contract", () => {
       let context: LSP8CompatibleERC721TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -171,7 +171,7 @@ describe("LSP8CompatibleERC721", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP8CompatibleERC721TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -190,8 +190,6 @@ describe("LSP8CompatibleERC721", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
