+++ tests/LSP8IdentifiableDigitalAsset/extensions/LSP8Mintable.test.ts
@@ -37,7 +37,7 @@ describe("LSP8Mintable", () => {
     describe("when deploying the contract", () => {
       let context: LSP8MintableTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -110,9 +110,10 @@ describe("LSP8Mintable", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP8MintableTestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
+
       describe("when initializing the contract", () => {
         shouldInitializeLikeLSP8(async () => {
           const { lsp8Mintable: lsp8, deployParams } = context;
@@ -128,8 +129,6 @@ describe("LSP8Mintable", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
@@ -141,7 +140,6 @@ describe("LSP8Mintable", () => {
       shouldBehaveLikeLSP8Mintable(() =>
         buildTestContext().then(async (context) => {
           await initializeProxy(context);
-
           return context;
         })
       );
