+++ tests/LSP9Vault/LSP9Vault.test.ts
@@ -118,7 +118,7 @@ describe("LSP9Vault", () => {
     describe("when deploying the contract", () => {
       let context: LSP9TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -222,7 +222,7 @@ describe("LSP9Vault", () => {
     describe("when deploying the contract as proxy", () => {
       let context: LSP9TestContext;
 
-      beforeEach(async () => {
+      before(async () => {
         context = await buildTestContext();
       });
 
@@ -241,8 +241,6 @@ describe("LSP9Vault", () => {
 
       describe("when calling initialize more than once", () => {
         it("should revert", async () => {
-          await initializeProxy(context);
-
           await expect(initializeProxy(context)).to.be.revertedWith(
             "Initializable: contract is already initialized"
           );
